commands: move snapshot command to separate pkg
