+++ command/snapshot/snapshot_command_test.go
@@ -1,0 +1,13 @@
+package snapshot
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestSnapshotCommand_noTabs(t *testing.T) {
+	t.Parallel()
+	if strings.ContainsRune(New().Help(), '\t') {
+		t.Fatal("usage has tabs")
+	}
+}
